+++ lib/datagrid/core.rb
@@ -4,27 +4,29 @@ require "active_support/core_ext/class/attribute"
 module Datagrid
   module Core
 
+    # @!visibility private
     def self.included(base)
-      base.extend         ClassMethods
+      base.extend ClassMethods
       base.class_eval do
         class_attribute :scope_value
 
         class_attribute :datagrid_attributes, instance_writer: false
         self.datagrid_attributes = []
 
-        class_attribute :dynamic_block, :instance_writer => false
+        class_attribute :dynamic_block, instance_writer: false
         class_attribute :forbidden_attributes_protection, instance_writer: false
         self.forbidden_attributes_protection = false
         if defined?(::ActiveModel::AttributeAssignment)
           include ::ActiveModel::AttributeAssignment
         end
       end
-      base.send :include, InstanceMethods
-    end # self.included
+      base.include InstanceMethods
+    end
 
     module ClassMethods
 
-      def datagrid_attribute(name, &block) #:nodoc:
+      # @!visibility private
+      def datagrid_attribute(name, &block)
         unless datagrid_attributes.include?(name)
           block ||= lambda do |value|
             value
@@ -41,6 +43,11 @@ module Datagrid
       end
 
       # Defines a scope at class level
+      # @return [void]
+      # @example
+      #   scope { User }
+      #   scope { Project.where(deleted: false) }
+      #   scope { Project.preload(:stages) }
       def scope(&block)
         if block
           current_scope = scope_value
@@ -49,17 +56,28 @@ module Datagrid
           }
           self
         else
-          check_scope_defined!
-          scope_value.call
+          scope = original_scope
+          driver.to_scope(scope)
         end
       end
 
-      def driver #:nodoc:
-        @driver ||= Drivers::AbstractDriver.guess_driver(scope).new
+      # @!visibility private
+      def original_scope
+        check_scope_defined!
+        scope_value.call
+      end
+
+      # @!visibility private
+      def driver
+        @driver ||= Drivers::AbstractDriver.guess_driver(scope_value.call).new
       end
 
       # Allows dynamic columns definition, that could not be defined at class level
-      #
+      # Columns that depend on the database state or third party service
+      # can be defined this way.
+      # @param block [Proc] block that defines dynamic columns
+      # @return [void]
+      # @example
       #   class MerchantsGrid
       #
       #     scope { Merchant }
@@ -69,11 +87,14 @@ module Datagrid
       #     dynamic do
       #       PurchaseCategory.all.each do |category|
       #         column(:"#{category.name.underscore}_sales") do |merchant|
-      #           merchant.purchases.where(:category_id => category.id).count
+      #           merchant.purchases.where(category_id: category.id).count
       #         end
       #       end
       #     end
       #   end
+      #
+      #   ProductCategory.create!(name: 'Swimwear')
+      #   ProductCategory.create!(name: 'Sportswear')
       #
       #   grid = MerchantsGrid.new
       #   grid.data # => [
@@ -95,6 +116,7 @@ module Datagrid
       end
 
       protected
+
       def check_scope_defined!(message = nil)
         message ||= "#{self}.scope is not defined"
         raise(Datagrid::ConfigurationError, message) unless scope_value
@@ -105,7 +127,7 @@ module Datagrid
         child_class.datagrid_attributes = self.datagrid_attributes.clone
       end
 
-    end # ClassMethods
+    end
 
     module InstanceMethods
 
@@ -122,8 +144,7 @@ module Datagrid
         end
       end
 
-      # Returns a hash of grid attributes including filter values
-      # and ordering values
+      # @return [Hash<Symbol, Object>] grid attributes including filter values and ordering values
       def attributes
         result = {}
         self.datagrid_attributes.each do |name|
@@ -132,20 +153,23 @@ module Datagrid
         result
       end
 
-      # Alias for <tt>send</tt> method
+      # @return [Object] Any datagrid attribute value
       def [](attribute)
         self.send(attribute)
       end
 
+      # Assigns any datagrid attribute
+      # @param attribute [Symbol, String] Datagrid attribute name
+      # @param value [Object] Datagrid attribute value
+      # @return [void]
       def []=(attribute, value)
         self.send(:"#{attribute}=", value)
       end
 
-      # Returns a scope(e.g ActiveRecord::Relation) with all applied filters
+      # @return [Object] a scope relation (e.g ActiveRecord::Relation) with all applied filters
       def assets
-        driver.to_scope(scope)
+        scope
       end
-
 
       # Updates datagrid attributes with a passed hash argument
       def attributes=(attributes)
@@ -163,7 +187,7 @@ module Datagrid
       end
 
       # Returns serializable query arguments skipping all nil values
-      #
+      # @example
       #   grid = ProductsGrid.new(category: 'dresses', available: true)
       #   grid.as_query # => {category: 'dresses', available: true}
       def as_query
@@ -174,8 +198,8 @@ module Datagrid
         attributes
       end
 
-      # Returns query parameters to link this grid from a page
-      #
+      # @return [Hash<Symbol, Hash<Symbol, Object>>] query parameters to link this grid from a page
+      # @example
       #   grid = ProductsGrid.new(category: 'dresses', available: true)
       #   Rails.application.routes.url_helpers.products_path(grid.query_params)
       #     # => "/products?products_grid[category]=dresses&products_grid[available]=true"
@@ -184,19 +208,18 @@ module Datagrid
       end
 
       # Redefines scope at instance level
-      #
+      # @example
       #   class MyGrid
       #     scope { Article.order('created_at desc') }
       #   end
       #
       #   grid = MyGrid.new
       #   grid.scope do |scope|
-      #     scope.where(:author_id => current_user.id)
+      #     scope.where(author_id: current_user.id)
       #   end
       #   grid.assets
       #       # => SELECT * FROM articles WHERE author_id = ?
       #       #    ORDER BY created_at desc
-      #
       def scope(&block)
         if block_given?
           current_scope = scope
@@ -205,29 +228,39 @@ module Datagrid
           }
           self
         else
-          check_scope_defined!
-          scope_value.call
+          scope = original_scope
+          driver.to_scope(scope)
         end
       end
 
+      # @!visibility private
+      def original_scope
+        check_scope_defined!
+        scope_value.call
+      end
+
       # Resets current instance scope to default scope defined in a class
+      # @return [void]
       def reset_scope
         self.scope_value = self.class.scope_value
       end
 
-      # Returns true if the scope was redefined for this instance of grid object
+      # @return [Boolean] true if the scope was redefined for this instance of grid object
       def redefined_scope?
         self.class.scope_value != scope_value
       end
 
-      def driver #:nodoc:
+      # @!visibility private
+      def driver
         self.class.driver
       end
 
-      def check_scope_defined!(message = nil) #:nodoc:
+      # @!visibility private
+      def check_scope_defined!(message = nil)
         self.class.send :check_scope_defined!, message
       end
 
+      # @return [String] a datagrid attributes and their values in inspection form
       def inspect
         attrs = attributes.map do |key, value|
           "#{key}: #{value.inspect}"
@@ -240,6 +273,6 @@ module Datagrid
           attributes == other.attributes &&
           scope == other.scope
       end
-    end # InstanceMethods
+    end
   end
 end